Phase-5/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend-phase5"
version = "0.1.0"
description = "Advanced Todo Backend with Event-Driven Architecture"
readme = "README.md"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.27",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "dapr>=1.11.0",
    "dapr-ext-fastapi>=1.11.0",
    "asyncpg>=0.31.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
    "confluent-kafka>=2.3.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "email-validator>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt==2.8.0",
    "python-jose[cryptography]>=3.5.0",
    "python-slugify>=8.0.0",
    "Pillow>=10.1.0",
    "httpx>=0.25.0",
    "openai-chatkit",
    "openai-agents",
    "sqlalchemy>=2.0.45",
    "python-dotenv",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.5.0",
    "faker>=19.6.0",
]

[tool.setuptools.packages.find]
where = ["backend/src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
addopts = "-v --cov=backend/src --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["backend/src"]
omit = ["*/tests/*", "*/conftest.py"]